+++ src/commands/index.ts
@@ -1,0 +1,7 @@
+import openElement from './openElement';
+import openUri from './openUri';
+
+export {
+  openElement,
+  openUri
+};