+++ src/languages/index.ts
@@ -1,0 +1,3 @@
+import sqCompletionItemProvider from './sqCompletionItem';
+
+export default sqCompletionItemProvider;