+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+### 20250509
+* 增加 `eslint`
+* 增加 `command`: `extension.openUri`
+
 ### 20240228
 * 增加 `commands`: `extension.openElement`，可在 VSCode 中打开 element-ui 官网